requirements.txt
Package            Version
------------------ ---------
brewer2mpl         1.4.1
certifi            2021.5.30
charset-normalizer 2.0.4
cycler             0.10.0
decorator          4.4.2
fonttools          4.33.3
idna               3.2
imageio            2.9.0
joblib             1.0.1
jsonpatch          1.32
jsonpointer        2.1
kiwisolver         1.3.1
lmdb               1.2.1
matplotlib         3.5.2
networkx           2.5.1
numpy              1.20.3
opencv-python      4.5.3.56
packaging          21.3
pandas             1.2.4
Pillow             8.2.0
pip                22.1.2
pyparsing          2.4.7
python-dateutil    2.8.1
pytz               2021.1
PyWavelets         1.1.1
pyzmq              22.2.1
requests           2.26.0
scikit-image       0.18.1
scikit-learn       0.24.2
scipy              1.8.1
seaborn            0.11.1
setuptools         41.2.0
six                1.16.0
threadpoolctl      2.1.0
tifffile           2021.4.8
torch              1.8.1
torchfile          0.1.0
torchsummary       1.5.1
torchvision        0.9.1
tornado            6.1
tqdm               4.61.0
typing-extensions  3.10.0.0
urllib3            1.26.6
visdom             0.1.8.9
websocket-client   1.1.0
z3-solver          4.8.12.0
